backend/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
defusedxml==0.5.0
Django==1.11
django-braces==1.12.0
django-cors-headers==2.1.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.7
idna==2.6
oauthlib==2.0.6
PyJWT==1.5.3
python-dotenv==0.7.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
